fix(ui): satisfy lint and deadcode gates for observer surfaces
