+++ ui/src/pages/chat/chat-observer-hud.test.ts
@@ -1,5 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import type { SessionObserverDigest } from "../../../../packages/gateway-protocol/src/schema/sessions.js";
+import { resolveChatPaneObserverRunId } from "../../lib/observer-digest.ts";
 import { createStorageMock } from "../../test-helpers/storage.ts";
 import { ChatObserverHudState, type ObserverHudInput } from "./components/chat-observer-hud.ts";
 
@@ -78,6 +79,34 @@ describe("ChatObserverHudState", () => {
         input({ running: false, activeRunId: null, digest: finalDigest, lastReadAt: 2_000 }),
       ),
     ).toBe("hidden");
+  });
+});
+
+describe("observer hud run identity from row data", () => {
+  it("shows a projected digest when attaching to an already-running session", () => {
+    const digest = {
+      sessionKey: "agent:main:current",
+      runId: "server-run",
+      revision: 1,
+      updatedAt: 2_000,
+      headline: "Already running",
+      health: "on-track" as const,
+    };
+    const activeRunId = resolveChatPaneObserverRunId({
+      localRunId: null,
+      session: { hasActiveRun: true, activeRunIds: ["server-run"] },
+      digest,
+    });
+
+    expect(activeRunId).toBe("server-run");
+    expect(
+      new ChatObserverHudState(false).mode({
+        running: activeRunId !== null,
+        activeRunId,
+        digest,
+        sideChatOpen: false,
+      }),
+    ).toBe("pill");
   });
 });
